+++ src/Bundle/DependencyInjection/SyliusSettingsExtension.php
@@ -21,11 +21,11 @@ use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
 final class SyliusSettingsExtension extends AbstractResourceExtension
 {
     /**
-     * {@inheritdoc}
+     * @throws \Exception
      */
-    public function load(array $config, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
-        $config = $this->processConfiguration($this->getConfiguration($config, $container), $config);
+        $config = $this->processConfiguration($this->getConfiguration($configs, $container), $configs);
         $loader = new XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
 
         $this->registerResources('sylius', $config['driver'], $config['resources'], $container);